3단계 - 지하철 구간 관리 (#1107)
* fix: 2단계 리뷰 반영

- docs: properties 주석 다시 작성
- fix: jpa 사용을 위한 생성자에 접근 제어자를 protected로 변경
- fix: 중복된 클래스명인 StationLineRequest를 StationLineSteps로 변경
- fix: 인수 테스트의 가독성을 위한 StationLineSteps의 메서드를 한글로 변경

* docs: 3주차 미션 정리

* refactor: 기존 응답값으로부터 Location 속성에서 ID를 추출하는 메서드를 유틸리티 클래스로 분리

- test: 지하철 역에 대해 생성 요청을 하는 메서드 및 클래스 추가

- test: 상행역과 하행역 ID를 받아 StationLineRequest 객체를 생성 후 반환하는 메서드 추가

- test: 지하철 역 mock 데이터 추가

* test: 지하철 구간 등록 인수 테스트 추가

* fix: @Nested 어노테이션 사용 시 java.net.ConnectException(Connection refused) 예외 발생으로 인한 삭제

- fix: 지하철 역 생성 요청 메서드에서 요청 및 응답 log 제거

* docs: 모든 테스트 메서드 상단에 주석으로 인수 조건 작성

* fix: 기존 upStationId와 downStationId 필드의 데이터 타입을 int에서 Long으로 변경

- fix: StationLine 엔티티 클래스에서 List<StationSection> 타입을 갖는 필드 추가 및 1:N 관계 맵핑

* feat: 지하철 구간 등록 기능 구현

- feat: 기존 지하철 노선 엔티티에 하행역을 수정하는 메서드 추가

- feat: 전역 예외 핸들러와 예외 응답 DTO 객체 추가

* refactor: 지하철 구간 등록 요청 및 응답에서 도메인에 종속된 DTO 객체 사용

* fix: StationLine이 save될 때, StationSection이 저장되도록 기존 Service 로직 수정

* refactor: 지하철 구간 저장 비즈니스 로직 리팩토링

* refactor: 기존 StationSection에 선언된 메서드를 StationLine 클래스로 이동

* refactor: 지하철 노선 목록 조회 테스트 메서드 리팩토링

- test: StationSection Mock 데이터 추가

- refactor: 지하철 구간 생성 요청 메서드 분리

* docs: properties 주석 삭제

* docs: properties 주석 추가

* docs: 지하철 구간 제거 요구사항 수정

* test: RestAssured를 사용하는 메서드를 응답 상태코드에 대한 검증 진행 유무를 확인할 수 있는 메서드명으로 변경

* test: 지하철 구간 제거 인수 테스트 추가

* fix: 기존 지하철역 관련 테스트 클래스에 @DirtiesContext 어노테이션 추가

* feat: 지하철 구간 제거 기능 구현

- fix: StationLine 클래스의 sections 필드에 CascadeType.REMOVE 옵션 추가

- feat: StationSection 클래스에 StationLine을 인자로 받는 생성자 추가

* refactor: 지하철 구간 제거 테스트 메서드 리팩토링

* docs: ReadMe update

* fix: 기존 Service 클래스에 선언된 public 메서드 중 prefix가 'save' 인 메서드의 prefix를 'create'로 수정

* feat: 구간 관리 기능 추가로 인한 StationLineReponse DTO 객체에 모든 구간을 표시할 수 있는 List<Station> 필드 추가

* fix: Entity 객체에 명시된 @JsonIgnore 어노테이션 삭제

- fix: 불명확한 메서드명 수정

* docs: 기존 인수 테스트 상단 주석에 표현 문장 중 'ID가 1보다 작은 숫자'를 '존재하지 않는 역'으로 변경

* fix: DTO 객체에 선언된 Line을 업데이트하는 기존 메서드(비즈니스 로직)를 정적 팩토리 메서드로 변경

* refactor: 명확하지 않은 메서드명 수정 및 필드 간 포맷팅 수정

- docs: TODO 주석 삭제

- docs: properties에 작성된 주석이 UTF-8 코드 값 디코딩한 형태로 볼 수 있게 수정

* docs: 3주차 리뷰 정리

* fix: 사용하지 않는 상수 정리

- fix: Entity 객체에 명시된 @JsonIgnore 어노테이션 삭제

* refactor: 기존 List<StationSection> 필드를 일급 컬렉션으로 변경 및 비즈니스 로직을 서비스에서 일급 컬렉션으로 이동

- docs: ReadMe update

* fix: 모든 구간 삭제 인수 테스트에서 노선이 가지는 구간 값을 검증하는 로직 추가

* docs: Update properties

* fix: StationLine Mock Data 수정

* test: 기존 Station Line의 삭제 관련 인수 테스트 보강

* feat: 구간 삭제 메서드를 StationSections(일급 컬렉션)으로 이동

- fix: 클래스 구현 순서에 맞춘 메서드 위치 변경

- docs: ReadMe update

* refactor: 영속성 전이(CASCADE) 옵션을 통한 부모/자식 엔티티 영속화

* test: 테스트 클래스 전용 Custom Annotation(@SpringBootTest, @Sql) 추가

- docs: 테스트 클래스에서 사용할 데이터베이스 script 추가

* refactor: 테스트 클래스 전용 Custom Annotation(@SpringBootTest, @Sql) 어노테이션 적용

* refactor: 기존 테스트 디렉토리 리팩토링

* test: @Nested 어노테이션을 활용한 테스트 메서드 분류

- docs: ReadMe update

* test: 기존 Sections 도메인 객체에 대한 테스트 코드에 @Nested 어노테이션 적용

- feat: Sections 도메인 객체에 마지막 역을 반환하는 메서드 추가

- test: Line과 Section Mock 데이터 추가

* docs: ReadMe update

+++ src/main/java/subway/dto/StationLineRequest.java
@@ -2,12 +2,16 @@ package subway.dto;
 
 public class StationLineRequest {
     private final String name;
+
     private final String color;
-    private final int upStationId;
-    private final int downStationId;
+
+    private final Long upStationId;
+
+    private final Long downStationId;
+
     private final int distance;
 
-    public StationLineRequest(String name, String color, int upStationId, int downStationId, int distance) {
+    public StationLineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -23,11 +27,11 @@ public class StationLineRequest {
         return color;
     }
 
-    public int getUpStationId() {
+    public Long getUpStationId() {
         return upStationId;
     }
 
-    public int getDownStationId() {
+    public Long getDownStationId() {
         return downStationId;
     }
